requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.0.1
pydantic==2.3.0
python-docx==0.8.11
PyPDF2==3.0.1
pdf2image==1.16.3
spacy==3.6.1
sentence-transformers==2.2.2
numpy==1.25.2
opencv-python==4.8.0.76